feat: remove graphql related code
