+++ src/core/data.rs
@@ -1,17 +1,18 @@
+use std::collections::HashSet;
+use std::ops::{Div, Mul, Neg};
+
+use bigdecimal::BigDecimal;
+use chrono::{NaiveDate, NaiveDateTime};
+use indexmap::IndexSet;
+use itertools::Itertools;
+
 use crate::core::account::Account;
+use crate::core::AccountName;
 use crate::core::amount::Amount;
 use crate::core::ledger::LedgerErrorType;
 use crate::core::models::{Flag, SingleTotalPrice, StringOrAccount, ZhangString};
 use crate::core::utils::inventory::{AmountLotPair, Inventory, LotInfo};
 use crate::core::utils::multi_value_map::MultiValueMap;
-use crate::core::AccountName;
-use bigdecimal::BigDecimal;
-use chrono::{NaiveDate, NaiveDateTime};
-use indexmap::IndexSet;
-use itertools::Itertools;
-use std::collections::HashSet;
-use std::ops::{Div, Mul, Neg};
-use std::sync::Arc;
 
 pub type Meta = MultiValueMap<String, ZhangString>;
 
@@ -65,21 +66,6 @@ pub struct Commodity {
 pub enum Balance {
     BalanceCheck(BalanceCheck),
     BalancePad(BalancePad),
-}
-
-impl Balance {
-    pub(crate) fn account_name(&self) -> &str {
-        match self {
-            Balance::BalanceCheck(check) => &check.account.content,
-            Balance::BalancePad(pad) => &pad.account.content,
-        }
-    }
-    pub(crate) fn date(&self) -> NaiveDate {
-        match self {
-            Balance::BalanceCheck(check) => check.date.naive_date(),
-            Balance::BalancePad(pad) => pad.date.naive_date(),
-        }
-    }
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -137,7 +123,6 @@ impl Transaction {
     pub(crate) fn get_postings_inventory(&self) -> Result<Inventory, LedgerErrorType> {
         let mut inventory = Inventory {
             currencies: Default::default(),
-            prices: Arc::new(Default::default()),
         };
         for posting in self.txn_postings() {
             let amount = posting.infer_trade_amount()?;
@@ -211,7 +196,6 @@ impl<'a> TxnPosting<'a> {
                 1 => {
                     let mut inventory = Inventory {
                         currencies: Default::default(),
-                        prices: Arc::new(Default::default()),
                     };
                     for (trade_amount, lot) in trade_amount_postings {
                         if let Some(trade_amount) = trade_amount {
@@ -343,10 +327,11 @@ pub struct Comment {
 #[cfg(test)]
 mod test {
     mod transaction {
+        use indoc::indoc;
+
         use crate::core::ledger::Ledger;
         use crate::core::models::Directive;
         use crate::parse_zhang;
-        use indoc::indoc;
 
         #[tokio::test]
         async fn should_return_true_given_balanced_transaction() {
@@ -477,13 +462,14 @@ mod test {
         }
 
         mod txn_posting {
+            use bigdecimal::BigDecimal;
+            use indoc::indoc;
+
             use crate::core::amount::Amount;
             use crate::core::data::Transaction;
             use crate::core::models::Directive;
             use crate::core::utils::inventory::LotInfo;
             use crate::parse_zhang;
-            use bigdecimal::BigDecimal;
-            use indoc::indoc;
 
             fn get_first_posting(content: &str) -> Transaction {
                 let directive = parse_zhang(content, None).unwrap().pop().unwrap();
