+++ src/error.rs
@@ -1,4 +1,5 @@
 use std::path::{Path, PathBuf};
+
 use thiserror::Error;
 
 pub type ZhangResult<T> = Result<T, ZhangError>;
