+++ src/lib.rs
@@ -1,17 +1,14 @@
-pub mod error;
-pub mod to_file;
 
-pub(crate) mod utils;
-
-#[allow(clippy::upper_case_acronyms)]
-#[allow(clippy::type_complexity)]
-pub mod p;
 
 pub mod cli;
 pub mod core;
+pub mod error;
 pub mod exporter;
 pub mod importer;
+#[allow(clippy::upper_case_acronyms)]
+#[allow(clippy::type_complexity)]
+pub mod p;
 pub mod server;
-
 pub mod target;
+
 pub use p::parse_zhang;
